Modified

+++ routers/merchant.js
@@ -20,7 +20,7 @@ router.get('/', (req, res) => {
                db.query(sql, (error,result)=>{
                    if(error) {
                     console.log(error)
-                   respons(500,error,"Server Error",res)
+                   respons(404,error,"Server Not Found",res)
                    }else
                    //data hasil dari mysql
                    respons(200,result,"Get in merchant",res)
@@ -28,8 +28,9 @@ router.get('/', (req, res) => {
             return
         }
     }
-    res.status(401)
-    res.send("Unauthorized")
+    console.log('Unauthorized')
+    res.status(401).send('Unauthorized');
+    
 })
   
 
@@ -39,11 +40,22 @@ router.post('/upload',(req,res)=>{
     const sql = `INSERT INTO merchant (password, name, join_date, phone_number) 
     VALUES ('${password}', '${name}', '${join_date}', '${phone_number}')`
 
+    if (Object.entries (req.body).length !==4 ||
+    !("password" in req.body) ||
+    !("name" in req.body) ||
+    !("join_date" in req.body) ||
+    !("phone_number" in req.body)
+    ){
+   respons(400 ,"Bad Request ","Unauthorized",res)
+    return  
+    }
+    
+
     db.query(sql,(error,result)=>{
       //error
         if(error){
           console.error("Error executing query:",error)
-         return respons(501 ,error,"Not Implemented",res)
+         return respons(401 ,error,"Unauthorized",res)
         }
         
         //Success
@@ -65,8 +77,21 @@ router.post('/upload',(req,res)=>{
 router.put('/upload/edit/:id',(req,res)=>{
     const id = req.params.id;
     const {password,name,join_date,phone_number}=req.body
-    const sql = `UPDATE merchant SET password='${password}', name='${name}', join_date='${join_date}', phone_number='${phone_number}' WHERE id = ${id}`;
-   db.query(sql,(error,result)=>{
+    const sql = `UPDATE merchant SET password='${password}', name='${name}', join_date='${join_date}',
+                 phone_number='${phone_number}' WHERE id = ${id}`;
+   
+                 if (Object.entries (req.body).length !==4 ||
+                 !("password" in req.body) ||
+                 !("name" in req.body) ||
+                 !("join_date" in req.body) ||
+                 !("phone_number" in req.body)
+                 ){
+
+   respons(404,"Invalid","Not Found",res)
+    return  
+    }
+   
+    db.query(sql,(error,result)=>{
     if(error) 
     respons(404,"Invalid","Not Found",res)
         console.log(result)
@@ -82,22 +107,33 @@ router.put('/upload/edit/:id',(req,res)=>{
 })
 
 
-router.delete('/upload/delete/:id',(req,res)=>{
-    const {id} = req.body
-   const sql= `DELETE FROM merchant WHERE id = ${id}`;
+router.delete('/upload/delete/:id', (req, res) => {
+    const id = req.params.id
+  
 
+    if (!id || !("password" in req.body)) {
+      return res.status(404).json({ error: "ID or password not found" });
+    }  
+    
+    const sql = `DELETE FROM merchant WHERE id = ${id}`;
 
-   db.query(sql,(error,result)=>{
-    if(error) respons(500,"Invalid","Server Error",res)
-        console.log(`Deleted Success`)
+    
+    db.query(sql, (error, result) => {
+        if (error) {
+            return respons(500, "Invalid", "Server Error", res);
+        } 
 
-    if(result?.affectedRows){ 
+      if (result.affectedRows) { 
         const data = {
-            isSuccess:result.affectedRows,
+          isSuccess: result.affectedRows,
         }
-        respons(200,data,"Deleted Success",res)
-    }
-})   
-   })
+        respons(200, data, "Successfully Deleted", res);
+      } else {
+        respons(404, (`Error id: ${id}`), "Data not found", res);
+        console.log(result)
+      }
+    })   
+  });
+  
 
 module.exports = router
